% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_biplot.R
\name{draw_biplot}
\alias{draw_biplot}
\title{Easily perform PCA and draw a biplot of the results with a calibrated axis}
\usage{
draw_biplot(
  data,
  key_axis = "none",
  ticklab = NULL,
  project = TRUE,
  scale = TRUE,
  fix_sign = FALSE,
  points = TRUE,
  arrows = TRUE,
  mult = NULL,
  colors = c("grey40", "red")
)
}
\arguments{
\item{data}{data.frame or object that can be coerced to one. PCA will be performed on numeric columns. The first non-numeric column will be used to label PC scores (points).}

\item{key_axis}{character indicating the column name of the axis to calibrate. If none is specified, no axes will be calibrated.}

\item{ticklab}{vector of values indicating axis breaks and tick labels for calibrated axis. If not specified, five evenly spaced round values will be chosen by \code{pretty()}. Ignored if \code{key_axis} is not specified.}

\item{project}{logical indicating whether projection lines should be drawn from PC scores to PC vectors. Defaults to \code{TRUE}.}

\item{scale}{logical passed to the \code{scale.} parameter of \code{prcomp()}. Defaults to \code{TRUE} (in contrast to \code{prcomp()}).}

\item{fix_sign}{logical indicating whether the signs of the loadings and scores should be chosen so that the first element of each loading is non-negative. Defaults to \code{FALSE} (in contrast to \code{princomp()} from which this parameter name was borrowed. \code{prcomp()} does not have such an option.)}

\item{points}{logical indicating whether to plot principal component scores. Defaults to \code{TRUE}.}

\item{arrows}{logical indicating whether to plot principal component loading vectors. Defaults to \code{TRUE}.}

\item{mult}{(NULL) numeric value used to scale the length of the rotation vectors, in effect creating secondary axes with different scales than the primary axes, although the secondary axes are not shown. If not specified, the factor is set to the ratio of 75th percentile distance between the scores and the origin to the length of the longest vector.}

\item{colors}{= vector of two colors, the first for the non-calibrated vectors and the second for the calibrated axis (and vector). Defaults to \code{c("grey40", "red")}.}
}
\description{
takes a data frame, performs PCA on the numeric columns, and draws a biplot with a clearly labeled calibrated axis for one of the original numeric columns. The first non-numeric column is used to label PC scores (points).
}
\details{
This function was developed mainly for teaching and learning purposes. There are many options for drawing biplots. The main contributions of this one are ease of use and option to calibrate one of the axes. Calibration calculations are performed by \code{calibrate::calibrate()}.
}
\section{References}{


(to be added)
}

\examples{

swiss \%>\%
  tibble::rownames_to_column() \%>\%
    draw_biplot("Agriculture")



}
