% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redav-package.R
\docType{package}
\name{redav-package}
\alias{redav}
\alias{redav-package}
\title{redav: R for EDAV}
\description{
(Under development) Will provides functions, data, and templates to accompany data visualization courses.
}
\author{
\strong{Maintainer}: Joyce Robbins \email{jtr13@columba.edu}

}
\keyword{internal}
