% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_missing}
\alias{plot_missing}
\title{Create a missing value plot in the style of \code{extracat::visna()} (no longer available on CRAN)}
\usage{
plot_missing(x, percent = TRUE)
}
\arguments{
\item{x}{data.frame or object that can be coerced to one}

\item{percent}{logical if \code{TRUE}, \% of rows and columns are shown, if \code{FALSE}, actual numbers of rows and columns are shown (defaults to \code{TRUE})}
}
\description{
Create a missing value plot in the style of \code{extracat::visna()} (no longer available on CRAN)
}
